+++ src/hoi4_world/roles/repeat_focus.h
@@ -8,7 +8,7 @@
 #include <vector>
 
 #include "src/hoi4_world/focus_trees/focus.h"
-#include "src/hoi4_world/roles/requirements/trigger_base.h"
+#include "src/hoi4_world/roles/triggers/trigger_base.h"
 #include "src/hoi4_world/world/hoi4_world.h"
