+++ src/hoi4_world/roles/repeat_focus_importer.cpp
@@ -2,7 +2,7 @@
 
 #include <external/commonItems/ParserHelpers.h>
 
-#include "requirements/always_trigger.h"
+#include "src/hoi4_world/roles/triggers/always_trigger.h"
