+++ src/hoi4_world/roles/repeat_focus_importer.h
@@ -7,7 +7,7 @@
 
 #include "src/hoi4_world/focus_trees/focus_importer.h"
 #include "src/hoi4_world/roles/repeat_focus.h"
-#include "src/hoi4_world/roles/requirements/trigger_importer.h"
+#include "src/hoi4_world/roles/triggers/trigger_importer.h"
