+++ src/hoi4_world/roles/role_importer.cpp
@@ -3,8 +3,8 @@
 #include <external/commonItems/CommonRegexes.h>
 #include <external/commonItems/ParserHelpers.h>
 
-#include "src/hoi4_world/roles/requirements/always_trigger.h"
-#include "src/hoi4_world/roles/requirements/tag_trigger.h"
+#include "src/hoi4_world/roles/triggers/always_trigger.h"
+#include "src/hoi4_world/roles/triggers/tag_trigger.h"
 
 
 hoi4::RoleImporter::RoleImporter()
