+++ src/hoi4_world/roles/role_importer.h
@@ -10,7 +10,7 @@
 #include "src/hoi4_world/events/event_importer.h"
 #include "src/hoi4_world/focus_trees/focus_importer.h"
 #include "src/hoi4_world/roles/repeat_focus_importer.h"
-#include "src/hoi4_world/roles/requirements/trigger_importer.h"
+#include "src/hoi4_world/roles/triggers/trigger_importer.h"
 #include "src/hoi4_world/roles/role.h"
