+++ src/hoi4_world/roles/role_importer_tests.cpp
@@ -3,9 +3,9 @@
 
 #include <sstream>
 
-#include "src/hoi4_world/roles/requirements/always_trigger.h"
-#include "src/hoi4_world/roles/requirements/and_trigger.h"
-#include "src/hoi4_world/roles/requirements/tag_trigger.h"
+#include "src/hoi4_world/roles/triggers/always_trigger.h"
+#include "src/hoi4_world/roles/triggers/and_trigger.h"
+#include "src/hoi4_world/roles/triggers/tag_trigger.h"
 #include "src/hoi4_world/roles/role.h"
 #include "src/hoi4_world/roles/role_importer.h"
