+++ src/hoi4_world/roles/stories_creator_tests.cpp
@@ -3,8 +3,8 @@
 
 #include <sstream>
 
-#include "src/hoi4_world/roles/requirements/always_trigger.h"
-#include "src/hoi4_world/roles/requirements/tag_trigger.h"
+#include "src/hoi4_world/roles/triggers/always_trigger.h"
+#include "src/hoi4_world/roles/triggers/tag_trigger.h"
 #include "src/hoi4_world/roles/stories_creator.h"
