+++ src/hoi4_world/roles/triggers/always_trigger.h
@@ -2,7 +2,7 @@
 
 
 
-#include "src/hoi4_world/roles/requirements/trigger_base.h"
+#include "src/hoi4_world/roles/triggers/trigger_base.h"
 #include "src/support/polymorphic_equality.h"
