+++ src/hoi4_world/roles/triggers/always_trigger_tests.cpp
@@ -4,7 +4,7 @@
 #include <sstream>
 
 #include "src/hoi4_world/countries/hoi4_country.h"
-#include "src/hoi4_world/roles/requirements/always_trigger.h"
+#include "src/hoi4_world/roles/triggers/always_trigger.h"
 #include "src/hoi4_world/states/hoi4_state.h"
 #include "src/hoi4_world/world/hoi4_world.h"
