+++ src/hoi4_world/roles/triggers/scope.h
