+++ src/hoi4_world/roles/triggers/trigger_importer_tests.cpp
@@ -3,12 +3,12 @@
 
 #include <sstream>
 
-#include "src/hoi4_world/roles/requirements/always_trigger.h"
-#include "src/hoi4_world/roles/requirements/and_trigger.h"
-#include "src/hoi4_world/roles/requirements/not_trigger.h"
-#include "src/hoi4_world/roles/requirements/or_trigger.h"
-#include "src/hoi4_world/roles/requirements/tag_trigger.h"
-#include "src/hoi4_world/roles/requirements/trigger_importer.h"
+#include "src/hoi4_world/roles/triggers/always_trigger.h"
+#include "src/hoi4_world/roles/triggers/and_trigger.h"
+#include "src/hoi4_world/roles/triggers/not_trigger.h"
+#include "src/hoi4_world/roles/triggers/or_trigger.h"
+#include "src/hoi4_world/roles/triggers/tag_trigger.h"
+#include "src/hoi4_world/roles/triggers/trigger_importer.h"
